requirements.txt
# Core dependencies
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.65.0

# LLM and API clients
openai>=1.0.0
requests>=2.31.0

# Dataset and embeddings
datasets>=2.14.0
transformers>=4.30.0
torch>=2.0.0
sentence-transformers>=2.2.0

# Graph libraries
networkx>=3.0
scikit-learn>=1.3.0

# Evaluation
evaluate>=0.4.0

# Logging and utilities
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
